// projects/wa-motif-open-api/auth-access-control-service/src/lib/model/permission.ts
/**
 * Motif Auth Access Control Service API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 2.0.0
 * Contact: info@vipera.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Permission { 
    /**
     * Component name
     */
    component: string;
    /**
     * Action (core values are VIEW, EXECUTE, MODIFY, * but can be customized)
     */
    action: string;
    /**
     * Method name or *(wildcard)
     */
    target: string;
    /**
     * Permission description
     */
    description: string;
}